Changed Float to Integer

+++ src/main/java/cf/nathanpb/RustCrafto/Radiation.java
@@ -22,7 +22,7 @@ public class Radiation {
                             if(((Sign)bs).getLine(0).contains("[RUSTCRAFT]")){
                                 String[] s2 = ((Sign)bs).getLine(1).split("=");
                                 if(s2[0].contains("RAD")){
-                                    addRadiation(p, Float.valueOf(s2[1]));
+                                    addRadiation(p, Integer.valueOf(s2[1]));
                                 }
                             }
                         }
@@ -33,18 +33,18 @@ public class Radiation {
     }
     public static void addRadiation(Player p, float rad){
         ProjectMetadata p2 = new ProjectMetadata(p.getUniqueId().toString(), Core.PlayerProfileDatabase);
-        if(!p2.hasKey("RADIATION")) p2.put("RADIATION", 0.0);
-        p2.put("RADIATION", p2.get("RADIATION", Float.class)+rad);
+        if(!p2.hasKey("RADIATION")) p2.put("RADIATION", 0);
+        p2.put("RADIATION", p2.get("RADIATION", Integer.class)+rad);
     }
     public static void removeRadiation(Player p, float rad){
         ProjectMetadata p2 = new ProjectMetadata(p.getUniqueId().toString(), Core.PlayerProfileDatabase);
         if(!p2.hasKey("RADIATION")) p2.put("RADIATION", 0.0);
-        if(p2.get("RADIATION", Float.class)-rad < 0) return;
-        p2.put("RADIATION", p2.get("RADIATION", Float.class)-rad);
+        if(p2.get("RADIATION", Integer.class)-rad < 0) return;
+        p2.put("RADIATION", p2.get("RADIATION", Integer.class)-rad);
     }
-    public static float getRadiation(Player p){
+    public static Integer getRadiation(Player p){
         ProjectMetadata p2 = new ProjectMetadata(p.getUniqueId().toString(), Core.PlayerProfileDatabase);
         if(!p2.hasKey("RADIATION")) p2.put("RADIATION", 0.0);
-        return p2.get("RADIATION", Float.class);
+        return p2.get("RADIATION", Integer.class);
     }
 }
